requirements.txt
django
djangorestframework
django-cors-headers
django-environ
psycopg2
requests
django-filter
